Created AlarmReceiver. Updated updateWeather() to use AlarmManager & PendingIntent.
